requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
flask==2.2.2
flask-cors==3.0.10
flask-marshmallow==0.14.0
flask-rest-paginate==1.1.3
flask-restful==0.3.9
flask-restplus==0.13.0
flask-sqlalchemy==3.0.2
itsdangerous==2.1.2
jinja2==3.1.2
jsonschema==4.17.3
markupsafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
pyrsistent==0.19.3
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
sqlalchemy==1.4.46
werkzeug==2.2.2
